+++ packages/fabric/editor-core/src/editor/plugins/text-formatting/index.tsx
@@ -41,8 +41,14 @@ export interface TextFormattingOptions {
   disableCode?: boolean;
 }
 
-const textFormatting = (options: TextFormattingOptions = {}): EditorPlugin => ({
-  marks() {
+const textFormatting: EditorPlugin = {
+  marks({ allowTextFormatting, textFormatting }) {
+    const options = textFormatting
+      ? textFormatting
+      : allowTextFormatting === true || !allowTextFormatting
+        ? {}
+        : allowTextFormatting;
+
     return [
       { name: 'em', mark: em, rank: 200 },
       { name: 'strong', mark: strong, rank: 300 },
@@ -112,6 +118,6 @@ const textFormatting = (options: TextFormattingOptions = {}): EditorPlugin => ({
       </ButtonsGroup>
     );
   },
-});
+};
 
 export default textFormatting;